move merge_lines implementation to its own source files to avoid duplicate symbols
